БОТ КАТЕГОРИИ/requirements.txt
aiogram
openai
python-dotenv
aiohttp
aiosqlite
